enh: added an initial code base for a sparse pattern and a sparse 1D matrix
These types are handling a sparse matrix by having a type to contain the
sparse pattern, and a second type which contains the actual data.

To initialize a sparse matrix one may do:

  type(sparse_pattern_t) :: sp
  type(sparse_matrix_t) :: S

  ! Create a new sparse pattern, size 100x100,
  ! with initial empty 20 elements per row.
  call sp%init(100, 100, np = 20)

  do ir = 1, sp%rows()
     call sp%add(ir, column)
  end do

  ! This cleans-up the sparse matrix by removing unnecessary
  ! elements in the sparse pattern
  call sp%finalize()

  ! Create the sparse matrix
  ! It is important to only create the sparse matrix
  ! once the sparse pattern has completed.
  call S%init(sp)

  ! Now S has all its entries (currently undefined) in
  !   S%M
  ! While the sparse pattern is contained in
  !   S%sp

The sparse pattern implements the standard CSR matrix with
index pointers starting at 1 and with size rows + 1. Once
the sparse pattern has been finalized the final element sp%rptr(nr + 1)
is equal to the number of non-zero elements in the matrix.

Signed-off-by: Nick Papior <nickpapior@gmail.com>