enh: added MPI distributions
Implemented a simple distribution type, based on inheritance.

Currently two distributions are implemented:

- Cyclic distribution with the first processor having
  element 1, second processor having element 2, etc. in
  a cyclic manner.
- Block cyclic distribution. This distribution is exactly
  equivalent to ScaLAPACK (basically the used methods
  are adapted from ScaLAPACK).

Both should work independently, but they require a communicator
to be passed to determine the size and rank of each processer.

TODO: currently these implementations forces the use of the MPI
library. However, we should add preprocessor statements to leviate
this.

Signed-off-by: Nick Papior <nickpapior@gmail.com>