enh: first full restructuring of the basis_ac type
basis_ac_t is split into 3 types, the top level basis_ac_t
which contains
 - all sites (atomic centered species).
 - number of unique species (to not confuse we name this a "state"
   This should probably be revised. But it maintains the name used
   in pspio, so the name is borrowed from there.
 - look-up tables which are used to index global sites to states,
   and or global orbitals to state or local orbital

The state is a sub-type contained in the basis with information
about all atomic orbitals that stem from this state. It only
contains:
 - n_orbital (number of unique orbitals)
 - orb, list of unique orbitals on this state (orbital_ac_t)

The last type is the orbital_ac_t. This is the essential container
which contains basic information such as:

 - l quantum number
 - m quantum number
 - occ initial occupation
 - R the radial function (a pspio_meshfunc_t which is a spline interpolator)

We should now be in a position to actually calculate stuff!

Signed-off-by: Nick Papior <nickpapior@gmail.com>