maint: Updated documentation and some review
I have reviewed:
 - elsi_*
 - energy
 - force
 - geometry

with some details and updates.
Now the number of atoms in automatically decided if Numberofatoms
is not in the input file.
Also, if the cell is not defined it will be 30 Bohr larger than
the atomic coordinates box. I.e. a molecule with vacuum gab of
15 Bohr, per +/- direction.

Added a calculate to force_t to update the total force.
Added a display to force_t to print-out the total force (optionally
all different terms).

Added lots of Doxygen documentation (although basic).

Signed-off-by: Nick Papior <nickpapior@gmail.com>