bug: added KB to the AC part, and also fixed GridXC calls
I know this commit is large and messy...

The KB terms are added to the AC part which was required.
I do find an inconsistency with regards to eigenvalues and
final energies comparing Siesta. However, it is running,
so any bugs found should easily propagate.

Secondly, there were a bug in using GridXC since it was
using the lower-level (cellXC) call. All GridXC routines
are exposed with gridxc_* names.
Now xc.F90 is fixed to use the correct interfaces and I have
also made sure it actually calculates some things.

Other changes:

- Minor space issues all-over
- Added a method in species.F90 to retrieve the r-max of
  a projector.
- Changed the calculation to use a periodic grid!
  This is what Siesta is doing (in terms of Hartree potential)
  so it should be easier to compare with this.

Fixes #26

Signed-off-by: Nick Papior <nickpapior@gmail.com>